Reorganized App and components
