+++ App/Home/Categories/IndividualCategories/categoryScreens.js
@@ -1,6 +1,6 @@
 import "react-native-gesture-handler";
 import React from "react";
-import { StyleSheet, Text, View, Button } from "react-native";
+import { Text, View, Button } from "react-native";
 
 import { RandomActivityScreen } from "./RandomActivityScreen";
